+++ sdk/core/azure-core/CHANGELOG.md
@@ -20,6 +20,7 @@
 
 ### Bug Fixes
 
+- Fixed the Curl transport adapter connection pooling when setting options.
 - Fixed for setting up the default transport adapter.
 
 ## 1.0.0-beta.3 (2020-11-11)